+++ src/hdrimage.cpp
@@ -25,6 +25,9 @@
 
 
 using namespace std;
+using Imath::V2f;
+using Imath::V3f;
+using Imath::M33f;
 
 // local functions
 namespace
@@ -33,24 +36,112 @@ namespace
 const Color4 g_blackPixel(0,0,0,0);
 
 // create a vector containing the normalized values of a 1D Gaussian filter
-Array2Df horizontalGaussianKernel(float sigma, float truncate);
-int wrapCoord(int p, int maxP, HDRImage::BorderMode m);
-void bilinearGreen(HDRImage &raw, int offsetX, int offsetY);
-void PhelippeauGreen(HDRImage &raw, const nanogui::Vector2i & red_offset);
-void MalvarGreen(HDRImage &raw, int c, const nanogui::Vector2i & red_offset);
-void MalvarRedOrBlueAtGreen(HDRImage &raw, int c, const nanogui::Vector2i &red_offset, bool horizontal);
-void MalvarRedOrBlue(HDRImage &raw, int c1, int c2, const nanogui::Vector2i &red_offset);
-void bilinearRedBlue(HDRImage &raw, int c, const nanogui::Vector2i & red_offset);
-void greenBasedRorB(HDRImage &raw, int c, const nanogui::Vector2i &red_offset);
-inline float clamp2(float value, float mn, float mx);
-inline float clamp4(float value, float a, float b, float c, float d);
-inline float interpGreenH(const HDRImage &raw, int x, int y);
-inline float interpGreenV(const HDRImage &raw, int x, int y);
-inline float ghG(const Array2Df & G, int i, int j);
-inline float gvG(const Array2Df & G, int i, int j);
-inline int bayerColor(int x, int y);
-inline nanogui::Vector3f cameraToLab(const nanogui::Vector3f c, const nanogui::Matrix3f & cameraToXYZ, const vector<float> & LUT);
+Array2Df horizontal_gaussian_kernel(float sigma, float truncate);
+int wrap_coord(int p, int maxP, HDRImage::BorderMode m);
+
+inline void clip_roi(Box2i & roi, const HDRImage & img)
+{
+    if (roi.has_volume())
+        roi.intersect(img.box());
+    else
+        roi = img.box();
+}
+
 } // namespace
+
+
+
+// makes a copy of the image, applies func for each pixel in roi of the copy, and returns the result
+HDRImage HDRImage::apply_function(function<Color4(const Color4 &)> func, Box2i roi) const
+{
+    clip_roi(roi, *this);
+
+    HDRImage result = (*this);
+
+    // for (int y = roi.min.y(); y < roi.max.y(); ++y)
+    parallel_for(roi.min.y(), roi.max.y(), [this,&func,&roi,&result](int y)
+    {
+        for (int x = roi.min.x(); x < roi.max.x(); ++x)
+            result(x,y) = func((*this)(x,y));
+    });
+
+    return result;
+}
+
+
+// applies func for each pixel in roi, and stores result back in this
+HDRImage & HDRImage::apply_function(function<Color4(const Color4 &)> func, Box2i roi)
+{
+    clip_roi(roi, *this);
+
+    // for (int y = roi.min.y(); y < roi.max.y(); ++y)
+    parallel_for(roi.min.y(), roi.max.y(), [this,&func,&roi](int y)
+    {
+        for (int x = roi.min.x(); x < roi.max.x(); ++x)
+            (*this)(x,y) = func((*this)(x,y));
+    });
+
+    return *this;
+}
+
+
+// for images this and other, computes result = func(this, other) for each pixel in roi, and returns the result
+// useful for something like c = a+b
+HDRImage HDRImage::apply_function(const HDRImage & other, function<Color4(const Color4 &, const Color4 &)> func, Box2i roi) const
+{
+    clip_roi(roi, *this);
+
+    HDRImage result = (*this);
+
+    // for (int y = roi.min.y(); y < roi.max.y(); ++y)
+    parallel_for(roi.min.y(), roi.max.y(), [this,&func,&roi,&result,&other](int y)
+    {
+        for (int x = roi.min.x(); x < roi.max.x(); ++x)
+            result(x,y) = func((*this)(x,y), other(x,y));
+    });
+
+    return result;
+}
+
+// for images this and other, computes this = func(this, other) for each pixel in roi, and stores result back in this
+// useful for something like a += b
+HDRImage & HDRImage::apply_function(const HDRImage & other, function<Color4(const Color4 &, const Color4 &)> func, Box2i roi)
+{
+    if (roi.has_volume())
+        roi.intersect(box());
+    else
+    {
+        roi = box();
+        roi.intersect(other.box());
+    }
+
+    // for (int y = roi.min.y(); y < roi.max.y(); ++y)
+    parallel_for(roi.min.y(), roi.max.y(), [this,&func,&roi,&other](int y)
+    {
+        for (int x = roi.min.x(); x < roi.max.x(); ++x)
+            (*this)(x,y) = func((*this)(x,y), other(x,y));
+    });
+
+    return *this;
+}
+
+
+Color4 HDRImage::reduce(function<Color4(const Color4 &, const Color4 &)> func, Box2i roi) const
+{
+    clip_roi(roi, *this);
+
+    Color4 result = (*this)(roi.min.x(),roi.min.y());
+
+    // for (int y = roi.min.y(); y < roi.max.y(); ++y)
+    parallel_for(roi.min.y(), roi.max.y(), [this,&func,&roi,&result](int y)
+    {
+        for (int x = roi.min.x(); x < roi.max.x(); ++x)
+            if (x != roi.min.x() && y != roi.min.y())
+                result = func(result, (*this)(x,y));
+    });
+
+    return result;
+}
 
 
 void HDRImage::copy_subimage(const HDRImage & src, Box2i roi, int dst_x, int dst_y)
@@ -103,8 +194,8 @@ const vector<string> & HDRImage::sampler_names()
 
 const Color4 & HDRImage::pixel(int x, int y, BorderMode mX, BorderMode mY) const
 {
-	x = wrapCoord(x, width(), mX);
-	y = wrapCoord(y, height(), mY);
+	x = wrap_coord(x, width(), mX);
+	y = wrap_coord(y, height(), mY);
 	if (x < 0 || y < 0)
 		return g_blackPixel;
 
@@ -113,8 +204,8 @@ const Color4 & HDRImage::pixel(int x, int y, BorderMode mX, BorderMode mY) const
 
 Color4 & HDRImage::pixel(int x, int y, BorderMode mX, BorderMode mY)
 {
-	x = wrapCoord(x, width(), mX);
-	y = wrapCoord(y, height(), mY);
+	x = wrap_coord(x, width(), mX);
+	y = wrap_coord(y, height(), mY);
 	if (x < 0 || y < 0)
 		throw out_of_range("Cannot assign to out-of-bounds pixel when BorderMode==BLACK.");
 
@@ -276,12 +367,12 @@ HDRImage HDRImage::convolved(const Array2Df &kernel, AtomicProgress progress,
 
 HDRImage HDRImage::gaussian_blurred_x(float sigmaX, AtomicProgress progress, BorderMode mX, float truncateX, Box2i roi) const
 {
-    return convolved(horizontalGaussianKernel(sigmaX, truncateX), progress, mX, mX, roi);
+    return convolved(horizontal_gaussian_kernel(sigmaX, truncateX), progress, mX, mX, roi);
 }
 
 HDRImage HDRImage::gaussian_blurred_y(float sigmaY, AtomicProgress progress, BorderMode mY, float truncateY, Box2i roi) const
 {
-    return convolved(horizontalGaussianKernel(sigmaY, truncateY).swapped_dims(), progress, mY, mY, roi);
+    return convolved(horizontal_gaussian_kernel(sigmaY, truncateY).swapped_dims(), progress, mY, mY, roi);
 }
 
 // Use principles of separability to blur an image using 2 1D Gaussian Filters
@@ -677,11 +768,7 @@ HDRImage HDRImage::resized_canvas(int newW, int newH, CanvasAnchor anchor, const
 
     nanogui::Vector2i bs(std::min(oldW, newW), std::min(oldH, newH));
 
-    // FIXME
-    // img.block(tlDst.x(), tlDst.y(),
-    //           bs.x(), bs.y()) = block(tlSrc.x(), tlSrc.y(),
-    //                                                 bs.x(), bs.y());
-    // img.copy_subimage();
+    img.copy_subimage(*this, Box2i(tlSrc, tlSrc+bs), tlDst.x(), tlDst.y());
     return img;
 }
 
@@ -697,377 +784,6 @@ HDRImage HDRImage::resized(int w, int h) const
     return newImage;
 }
 
-/*!
- * \brief Multiplies a raw image by the Bayer mosaic pattern so that only a single
- * R, G, or B channel is non-zero for each pixel.
- *
- * We assume the canonical Bayer pattern looks like:
- *
- * \rst
- * +---+---+
- * | R | G |
- * +---+---+
- * | G | B |
- * +---+---+
- *
- * \endrst
- *
- * and the pattern is tiled across the entire image.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern
- */
-void HDRImage::bayer_mosaic(const nanogui::Vector2i &red_offset)
-{
-    Color4 mosaic[2][2] = {{Color4(1.f, 0.f, 0.f, 1.f), Color4(0.f, 1.f, 0.f, 1.f)},
-                           {Color4(0.f, 1.f, 0.f, 1.f), Color4(0.f, 0.f, 1.f, 1.f)}};
-    for (int y = 0; y < height(); ++y)
-    {
-        int r = mod(y - red_offset.y(), 2);
-        for (int x = 0; x < width(); ++x)
-        {
-            int c = mod(x - red_offset.x(), 2);
-            (*this)(x,y) *= mosaic[r][c];
-        }
-    }
-}
-
-
-/*!
- * \brief Compute the missing green pixels using a simple bilinear interpolation.
- * from the 4 neighbors.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_green_linear(const nanogui::Vector2i &red_offset)
-{
-    bilinearGreen(*this, red_offset.x(), red_offset.y());
-}
-
-/*!
- * \brief Compute the missing green pixels using vertical linear interpolation.
- *
- * @param raw       The source raw pixel data.
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_green_horizontal(const HDRImage &raw, const nanogui::Vector2i &red_offset)
-{
-    parallel_for(red_offset.y(), height(), 2, [this,&raw,&red_offset](int y)
-    {
-        for (int x = 2+red_offset.x(); x < width()-2; x += 2)
-        {
-            // populate the green channel into the red and blue pixels
-            (*this)(x  , y  ).g = interpGreenH(raw, x, y);
-            (*this)(x+1, y+1).g = interpGreenH(raw, x + 1, y + 1);
-        }
-    });
-}
-
-/*!
- * \brief Compute the missing green pixels using horizontal linear interpolation.
- *
- * @param raw       The source raw pixel data.
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_green_vertical(const HDRImage &raw, const nanogui::Vector2i &red_offset)
-{
-    parallel_for(2+red_offset.y(), height()-2, 2, [this,&raw,&red_offset](int y)
-    {
-        for (int x = red_offset.x(); x < width(); x += 2)
-        {
-            (*this)(x  , y  ).g = interpGreenV(raw, x, y);
-            (*this)(x+1, y+1).g = interpGreenV(raw, x + 1, y + 1);
-        }
-    });
-}
-
-/*!
- * \brief Interpolate the missing green pixels using the method by Malvar et al. 2004.
- *
- * The method uses a plus "+" shaped 5x5 filter, which is linear, except--to reduce
- * ringing/over-shooting--the interpolation is not allowed to extrapolate higher or
- * lower than the surrounding green pixels.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_green_malvar(const nanogui::Vector2i &red_offset)
-{
-    // fill in missing green at red pixels
-    MalvarGreen(*this, 0, red_offset);
-    // fill in missing green at blue pixels
-    MalvarGreen(*this, 2, nanogui::Vector2i((red_offset.x() + 1) % 2, (red_offset.y() + 1) % 2));
-}
-
-/*!
- * \brief Interpolate the missing green pixels using the method by Phelippeau et al. 2009.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_green_phelippeau(const nanogui::Vector2i &red_offset)
-{
-    PhelippeauGreen(*this, red_offset);
-}
-
-/*!
- * \brief Interpolate the missing red and blue pixels using a simple linear or bilinear interpolation.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_red_blue_linear(const nanogui::Vector2i &red_offset)
-{
-    bilinearRedBlue(*this, 0, red_offset);
-    bilinearRedBlue(*this, 2, nanogui::Vector2i((red_offset.x() + 1) % 2, (red_offset.y() + 1) % 2));
-}
-
-/*!
- * \brief Interpolate the missing red and blue pixels using a linear or bilinear interpolation
- * guided by the green channel, which is assumed already demosaiced.
- *
- * The interpolation is equivalent to performing (bi)linear interpolation of the red-green and
- * blue-green differences, and then adding green back into the interpolated result. This inject
- * some of the higher resolution of the green channel, and reduces color fringing under the
- * assumption that the color channels in natural images are positively correlated.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_red_blue_green_guided_linear(const nanogui::Vector2i &red_offset)
-{
-    greenBasedRorB(*this, 0, red_offset);
-    greenBasedRorB(*this, 2, nanogui::Vector2i((red_offset.x() + 1) % 2, (red_offset.y() + 1) % 2));
-}
-
-/*!
- * \brief Interpolate the missing red and blue pixels using the method by Malvar et al. 2004.
- *
- * The interpolation for each channel is guided by the available information from all other
- * channels. The green channel is assumed to already be demosaiced.
- *
- * The method uses a 5x5 linear filter.
- *
- * @param red_offset The x,y offset to the first red pixel in the Bayer pattern.
- */
-void HDRImage::demosaic_red_blue_malvar(const nanogui::Vector2i &red_offset)
-{
-    // fill in missing red horizontally
-    MalvarRedOrBlueAtGreen(*this, 0, nanogui::Vector2i((red_offset.x() + 1) % 2, red_offset.y()), true);
-    // fill in missing red vertically
-    MalvarRedOrBlueAtGreen(*this, 0, nanogui::Vector2i(red_offset.x(), (red_offset.y() + 1) % 2), false);
-
-    // fill in missing blue horizontally
-    MalvarRedOrBlueAtGreen(*this, 2, nanogui::Vector2i(red_offset.x(), (red_offset.y() + 1) % 2), true);
-    // fill in missing blue vertically
-    MalvarRedOrBlueAtGreen(*this, 2, nanogui::Vector2i((red_offset.x() + 1) % 2, red_offset.y()), false);
-
-    // fill in missing red at blue
-    MalvarRedOrBlue(*this, 0, 2, nanogui::Vector2i((red_offset.x() + 1) % 2, (red_offset.y() + 1) % 2));
-    // fill in missing blue at red
-    MalvarRedOrBlue(*this, 2, 0, red_offset);
-}
-
-/*!
- * \brief Reduce some remaining color fringing and zipper artifacts by median-filtering the
- * red-green and blue-green differences as originally proposed by Freeman.
- */
-HDRImage HDRImage::median_filter_bayer_artifacts() const
-{
-    AtomicProgress progress;
-    HDRImage colorDiff = apply_function([](const Color4 & c){return Color4(c.r-c.g,c.g,c.b-c.g,c.a);});
-    colorDiff = colorDiff.median_filtered(1.f, 0, AtomicProgress(progress, .5f))
-                         .median_filtered(1.f, 2, AtomicProgress(progress, .5f));
-    return apply_function(colorDiff, [](const Color4 & i, const Color4 & med){return Color4(med.r + i.g, i.g, med.b + i.g, i.a);});
-}
-
-/*!
- * \brief Demosaic the image using the "Adaptive Homogeneity-Directed" interpolation
- * approach proposed by Hirakawa et al. 2004.
- *
- * The approach is fairly expensive, but produces the best results.
- *
- * The method first creates two competing full-demosaiced images: one where the
- * green channel is interpolated vertically, and the other horizontally. In both
- * images the red and green are demosaiced using the corresponding green channel
- * as a guide.
- *
- * The two candidate images are converted to XYZ (using the supplied \a cameraToXYZ
- * matrix) subsequently to CIE L*a*b* space in order to determine how perceptually
- * "homogeneous" each pixel neighborhood is.
- *
- * "Homogeneity maps" are created for the two candidate imates which count, for each
- * pixel, the number of perceptually similar pixels among the 4 neighbors in the
- * cardinal directions.
- *
- * Finally, the output image is formed by choosing for each pixel the demosaiced
- * result which has the most homogeneous "votes" in the surrounding 3x3 neighborhood.
- *
- * @param red_offset     The x,y offset to the first red pixel in the Bayer pattern.
- * @param cameraToXYZ   The matrix that transforms from sensor values to XYZ with
- *                      D65 white point.
- */
-void HDRImage::demosaicAHD(const nanogui::Vector2i &red_offset, const nanogui::Matrix3f &cameraToXYZ)
-{
-    using Image3f = Array2D<nanogui::Vector3f>;
-    using HomoMap = Array2D<uint8_t>;
-    HDRImage rgbH = *this;
-    HDRImage rgbV = *this;
-    Image3f labH(width(), height(), nanogui::Vector3f(0.f));
-    Image3f labV(width(), height(), nanogui::Vector3f(0.f));
-    HomoMap homoH = HomoMap(width(), height());
-    HomoMap homoV = HomoMap(width(), height());
-
-    // interpolate green channel both horizontally and vertically
-    rgbH.demosaic_green_horizontal(*this, red_offset);
-    rgbV.demosaic_green_vertical(*this, red_offset);
-
-    // interpolate the red and blue using the green as a guide
-    rgbH.demosaic_red_blue_green_guided_linear(red_offset);
-    rgbV.demosaic_red_blue_green_guided_linear(red_offset);
-
-    // Scale factor to push XYZ values to [0,1] range
-    // FIXME
-    float scale = 1.0f;// / (max().max() * cameraToXYZ.max());
-
-    // Precompute a table for the nonlinear part of the CIELab conversion
-    vector<float> labLUT;
-    labLUT.reserve(0xFFFF);
-    parallel_for(0, labLUT.size(), [&labLUT](int i)
-    {
-        float r = i * 1.0f / (labLUT.size()-1);
-        labLUT[i] = r > 0.008856 ? std::pow(r, 1.0f / 3.0f) : 7.787f*r + 4.0f/29.0f;
-    });
-
-    // convert both interpolated images to CIE L*a*b* so we can compute perceptual differences
-    parallel_for(0, height(), [&rgbH,&labH,&cameraToXYZ,&labLUT,scale](int y)
-    {
-        for (int x = 0; x < rgbH.width(); ++x)
-            labH(x,y) = cameraToLab(nanogui::Vector3f(rgbH(x,y)[0], rgbH(x,y)[1], rgbH(x,y)[2])*scale, cameraToXYZ, labLUT);
-    });
-    parallel_for(0, height(), [&rgbV,&labV,&cameraToXYZ,&labLUT,scale](int y)
-    {
-        for (int x = 0; x < rgbV.width(); ++x)
-            labV(x,y) = cameraToLab(nanogui::Vector3f(rgbV(x,y)[0], rgbV(x,y)[1], rgbV(x,y)[2])*scale, cameraToXYZ, labLUT);
-    });
-
-    // Build homogeneity maps from the CIELab images which count, for each pixel,
-    // the number of visually similar neighboring pixels
-    static const int neighbor[4][2] = { {-1, 0}, {1, 0}, {0, -1}, {0, 1} };
-    parallel_for(1, height()-1, [&homoH,&homoV,&labH,&labV](int y)
-    {
-        for (int x = 1; x < labH.width()-1; ++x)
-        {
-            float ldiffH[4], ldiffV[4], abdiffH[4], abdiffV[4];
-
-            for (int i = 0; i < 4; i++)
-            {
-                int dx = neighbor[i][0];
-                int dy = neighbor[i][1];
-
-                // Local luminance and chromaticity differences to the 4 neighbors for both interpolations directions
-                ldiffH[i] = std::abs(labH(x,y)[0] - labH(x+dx,y+dy)[0]);
-                ldiffV[i] = std::abs(labV(x,y)[0] - labV(x+dx,y+dy)[0]);
-                abdiffH[i] = ::square(labH(x,y)[1] - labH(x+dx,y+dy)[1]) +
-                             ::square(labH(x,y)[2] - labH(x+dx,y+dy)[2]);
-                abdiffV[i] = ::square(labV(x,y)[1] - labV(x+dx,y+dy)[1]) +
-                             ::square(labV(x,y)[2] - labV(x+dx,y+dy)[2]);
-            }
-
-            float leps = std::min(std::max(ldiffH[0], ldiffH[1]),
-                                  std::max(ldiffV[2], ldiffV[3]));
-            float abeps = std::min(std::max(abdiffH[0], abdiffH[1]),
-                                   std::max(abdiffV[2], abdiffV[3]));
-
-            // Count number of neighboring pixels that are visually similar
-            for (int i = 0; i < 4; i++)
-            {
-                if (ldiffH[i] <= leps && abdiffH[i] <= abeps)
-                    homoH(x,y)++;
-                if (ldiffV[i] <= leps && abdiffV[i] <= abeps)
-                    homoV(x,y)++;
-            }
-        }
-    });
-
-    // Combine the most homogenous pixels for the final result
-    parallel_for(1, height()-1, [this,&homoH,&homoV,&rgbH,&rgbV](int y)
-    {
-        for (int x = 1; x < this->width()-1; ++x)
-        {
-            // Sum up the homogeneity of both images in a 3x3 window
-            int hmH = 0, hmV = 0;
-            for (int j = y-1; j <= y+1; j++)
-                for (int i = x-1; i <= x+1; i++)
-                {
-                    hmH += homoH(i, j);
-                    hmV += homoV(i, j);
-                }
-
-            if (hmH > hmV)
-            {
-                // horizontal interpolation is more homogeneous
-                (*this)(x,y) = rgbH(x,y);
-            }
-            else if (hmV > hmH)
-            {
-                // vertical interpolation is more homogeneous
-                (*this)(x,y) = rgbV(x,y);
-            }
-            else
-            {
-                // No clear winner, blend
-                Color4 blend = (rgbH(x,y) + rgbV(x,y)) * 0.5f;
-                (*this)(x,y) = blend;
-            }
-        }
-    });
-
-    // Now handle the boundary pixels
-    demosaic_border(3);
-}
-
-/*!
- * \brief   Demosaic the border of the image using naive averaging.
- *
- * Provides a results for all border pixels using a straight averge of the available pixels
- * in the 3x3 neighborhood. Useful in combination with more sophisticated methods which
- * require a larger window, and therefore cannot produce results at the image boundary.
- *
- * @param border    The size of the border in pixels.
- */
-void HDRImage::demosaic_border(size_t border)
-{
-    parallel_for(0, height(), [&](size_t y)
-    {
-        for (size_t x = 0; x < (size_t)width(); ++x)
-        {
-            // skip the center of the image
-            if (x == border && y >= border && y < height() - border)
-                x = width() - border;
-
-            nanogui::Vector3f sum = nanogui::Vector3f(0.f);
-            nanogui::Vector3i count = nanogui::Vector3i(0.f);
-
-            for (size_t ys = y - 1; ys <= y + 1; ++ys)
-            {
-                for (size_t xs = x - 1; xs <= x + 1; ++xs)
-                {
-                    // rely on xs and ys = -1 to wrap around to max value
-                    // since they are unsigned
-                    if (ys < (size_t)height() && xs < (size_t)width())
-                    {
-                        int c = bayerColor(xs, ys);
-                        sum[c] += (*this)(xs,ys)[c];
-                        ++count[c];
-                    }
-                }
-            }
-
-            int col = bayerColor(x, y);
-            for (int c = 0; c < 3; ++c)
-            {
-                if (col != c)
-                    (*this)(x,y)[c] = count[c] ? (sum[c] / count[c]) : 1.0f;
-            }
-        }
-    });
-}
 
 /*!
  * @brief           Apply a global brightness+contrast adjustment to the RGB pixel values.
@@ -1182,7 +898,7 @@ namespace
 {
 
 // create a vector containing the normalized values of a 1D Gaussian filter
-Array2Df horizontalGaussianKernel(float sigma, float truncate)
+Array2Df horizontal_gaussian_kernel(float sigma, float truncate)
 {
     // calculate the size of the filter
     int offset = int(std::ceil(truncate * sigma));
@@ -1205,7 +921,7 @@ Array2Df horizontalGaussianKernel(float sigma, float truncate)
     return fData;
 }
 
-int wrapCoord(int p, int maxP, HDRImage::BorderMode m)
+int wrap_coord(int p, int maxP, HDRImage::BorderMode m)
 {
     if (p >= 0 && p < maxP)
         return p;
@@ -1225,247 +941,6 @@ int wrapCoord(int p, int maxP, HDRImage::BorderMode m)
         default:
             return -1;
     }
-}
-
-inline nanogui::Vector3f cameraToLab(const nanogui::Vector3f c, const nanogui::Matrix3f & cameraToXYZ, const vector<float> & LUT)
-{
-    nanogui::Vector3f xyz = cameraToXYZ * c;
-
-    for (int i = 0; i < 3; ++i)
-        xyz[i] = LUT[::clamp((int) (xyz[i] * LUT.size()), 0, int(LUT.size()-1))];
-
-    return nanogui::Vector3f(116.0f * xyz[1] - 16, 500.0f * (xyz[0] - xyz[1]), 200.0f * (xyz[1] - xyz[2]));
-}
-
-inline int bayerColor(int x, int y)
-{
-    const int bayer[2][2] = {{0,1},{1,2}};
-
-    return bayer[y % 2][x % 2];
-}
-
-inline float clamp2(float value, float mn, float mx)
-{
-    if (mn > mx)
-        std::swap(mn, mx);
-    return ::clamp(value, mn, mx);
-}
-
-inline float clamp4(float value, float a, float b, float c, float d)
-{
-    float mn = min(a, b, c, d);
-    float mx = max(a, b, c, d);
-    return ::clamp(value, mn, mx);
-}
-
-inline float interpGreenH(const HDRImage &raw, int x, int y)
-{
-    float v = 0.50f * (raw(x - 1, y).g + raw(x + 1, y).g + raw(x, y).g) -
-              0.25f * (raw(x - 2, y).g + raw(x + 2, y).g);
-    // Don't extrapolate past the neighboring green values
-    return clamp2(v, raw(x - 1, y).g, raw(x + 1, y).g);
-}
-
-inline float interpGreenV(const HDRImage &raw, int x, int y)
-{
-    float v = 0.50f * (raw(x, y - 1).g + raw(x, y + 1).g + raw(x, y).g) -
-              0.25f * (raw(x, y - 2).g + raw(x, y + 2).g);
-    // Don't extrapolate past the neighboring green values
-    return clamp2(v, raw(x, y - 1).g, raw(x, y + 1).g);
-}
-
-inline float ghG(const Array2Df & G, int i, int j)
-{
-    return fabs(G(i-1,j) - G(i,j)) + fabs(G(i+1,j) - G(i,j));
-}
-
-inline float gvG(const Array2Df & G, int i, int j)
-{
-    return fabs(G(i,j-1) - G(i,j)) + fabs(G(i,j+1) - G(i,j));
-}
-
-void bilinearGreen(HDRImage &raw, int offsetX, int offsetY)
-{
-    parallel_for(1, raw.height()-1-offsetY, 2, [&raw,offsetX,offsetY](int yy)
-    {
-        int t = yy + offsetY;
-        for (int xx = 1; xx < raw.width()-1-offsetX; xx += 2)
-        {
-            int l = xx + offsetX;
-
-            // coordinates of the missing green pixels (red and blue) in
-            // this Bayer tile are: (l,t) and (r,b)
-            int r = l+1;
-            int b = t+1;
-
-            raw(l, t).g = 0.25f * (raw(l, t - 1).g + raw(l, t + 1).g +
-                                   raw(l - 1, t).g + raw(l + 1, t).g);
-            raw(r, b).g = 0.25f * (raw(r, b - 1).g + raw(r, b + 1).g +
-                                   raw(r - 1, b).g + raw(r + 1, b).g);
-        }
-    });
-}
-
-
-void PhelippeauGreen(HDRImage &raw, const nanogui::Vector2i & red_offset)
-{
-    Array2Df Gh(raw.width(), raw.height());
-    Array2Df Gv(raw.width(), raw.height());
-
-    // populate horizontally interpolated green
-    parallel_for(red_offset.y(), raw.height(), 2, [&raw,&Gh,&red_offset](int y)
-    {
-        for (int x = 2+red_offset.x(); x < raw.width() - 2; x += 2)
-        {
-            Gh(x  , y  ) = interpGreenH(raw, x, y);
-            Gh(x+1, y+1) = interpGreenH(raw, x + 1, y + 1);
-        }
-    });
-
-    // populate vertically interpolated green
-    parallel_for(2+red_offset.y(), raw.height()-2, 2, [&raw,&Gv,&red_offset](int y)
-    {
-        for (int x = red_offset.x(); x < raw.width(); x += 2)
-        {
-            Gv(x  , y  ) = interpGreenV(raw, x, y);
-            Gv(x+1, y+1) = interpGreenV(raw, x + 1, y + 1);
-        }
-    });
-
-    parallel_for(2+red_offset.y(), raw.height()-2, 2, [&raw,&Gh,&Gv,red_offset](int y)
-    {
-        for (int x = 2+red_offset.x(); x < raw.width()-2; x += 2)
-        {
-            float ghGh = ghG(Gh, x, y);
-            float ghGv = ghG(Gv, x, y);
-            float gvGh = gvG(Gh, x, y);
-            float gvGv = gvG(Gv, x, y);
-
-            raw(x, y).g = (ghGh + gvGh <= gvGv + ghGv) ? Gh(x, y) : Gv(x, y);
-
-            x++;
-            y++;
-
-            ghGh = ghG(Gh, x, y);
-            ghGv = ghG(Gv, x, y);
-            gvGh = gvG(Gh, x, y);
-            gvGv = gvG(Gv, x, y);
-
-            raw(x, y).g = (ghGh + gvGh <= gvGv + ghGv) ? Gh(x, y) : Gv(x, y);
-        }
-    });
-}
-
-
-void MalvarGreen(HDRImage &raw, int c, const nanogui::Vector2i & red_offset)
-{
-    // fill in half of the missing locations (R or B)
-    parallel_for(2, raw.height()-2-red_offset.y(), 2, [&raw,c,&red_offset](int yy)
-    {
-        int y = yy + red_offset.y();
-        for (int xx = 2; xx < raw.width()-2-red_offset.x(); xx += 2)
-        {
-            int x = xx + red_offset.x();
-            float v = (4.f * raw(x,y)[c]
-                         + 2.f * (raw(x, y-1)[1] + raw(x-1, y)[1] +
-                                  raw(x, y+1)[1] + raw(x+1, y)[1])
-                         - 1.f * (raw(x, y-2)[c] + raw(x-2, y)[c] +
-                                  raw(x, y+2)[c] + raw(x+2, y)[c]))/8.f;
-            raw(x,y)[1] = clamp4(v, raw(x, y-1)[1], raw(x-1, y)[1],
-                                    raw(x, y+1)[1], raw(x+1, y)[1]);
-        }
-    });
-}
-
-void MalvarRedOrBlueAtGreen(HDRImage &raw, int c, const nanogui::Vector2i &red_offset, bool horizontal)
-{
-    int dx = (horizontal) ? 1 : 0;
-    int dy = (horizontal) ? 0 : 1;
-    // fill in half of the missing locations (R or B)
-    parallel_for(2+red_offset.y(), raw.height()-2, 2, [&raw,c,&red_offset,dx,dy](int y)
-    {
-        for (int x = 2+red_offset.x(); x < raw.width()-2; x += 2)
-        {
-            raw(x,y)[c] = (5.f * raw(x,y)[1]
-                         - 1.f * (raw(x-1,y-1)[1] + raw(x+1,y-1)[1] + raw(x+1,y+1)[1] + raw(x-1,y+1)[1] + raw(x-2,y)[1] + raw(x+2,y)[1])
-                         + .5f * (raw(x,y-2)[1] + raw(x,y+2)[1])
-                         + 4.f * (raw(x-dx,y-dy)[c] + raw(x+dx,y+dy)[c]))/8.f;
-        }
-    });
-}
-
-void MalvarRedOrBlue(HDRImage &raw, int c1, int c2, const nanogui::Vector2i &red_offset)
-{
-    // fill in half of the missing locations (R or B)
-    parallel_for(2+red_offset.y(), raw.height()-2, 2, [&raw,c1,c2,&red_offset](int y)
-    {
-        for (int x = 2+red_offset.x(); x < raw.width()-2; x += 2)
-        {
-            raw(x,y)[c1] = (6.f * raw(x,y)[c2]
-                          + 2.f * (raw(x-1,y-1)[c1] + raw(x+1,y-1)[c1] + raw(x+1,y+1)[c1] + raw(x-1,y+1)[c1])
-                          - 3/2.f * (raw(x,y-2)[c2] + raw(x,y+2)[c2] + raw(x-2,y)[c2] + raw(x+2,y)[c2]))/8.f;
-        }
-    });
-}
-
-
-// takes as input a raw image and returns a single-channel
-// 2D image corresponding to the red or blue channel using simple interpolation
-void bilinearRedBlue(HDRImage &raw, int c, const nanogui::Vector2i & red_offset)
-{
-    // diagonal interpolation
-    parallel_for(red_offset.y() + 1, raw.height()-1, 2, [&raw,c,&red_offset](int y)
-    {
-        for (int x = red_offset.x() + 1; x < raw.width() - 1; x += 2)
-            raw(x, y)[c] = 0.25f * (raw(x - 1, y - 1)[c] + raw(x + 1, y - 1)[c] +
-                                    raw(x - 1, y + 1)[c] + raw(x + 1, y + 1)[c]);
-    });
-
-    // horizontal interpolation
-    parallel_for(red_offset.y(), raw.height(), 2, [&raw,c,&red_offset](int y)
-    {
-        for (int x = red_offset.x() + 1; x < raw.width() - 1; x += 2)
-            raw(x, y)[c] = 0.5f * (raw(x - 1, y)[c] + raw(x + 1, y)[c]);
-    });
-
-    // vertical interpolation
-    parallel_for(red_offset.y() + 1, raw.height() - 1, 2, [&raw,c,&red_offset](int y)
-    {
-        for (int x = red_offset.x(); x < raw.width(); x += 2)
-            raw(x, y)[c] = 0.5f * (raw(x, y - 1)[c] + raw(x, y + 1)[c]);
-    });
-}
-
-
-// takes as input a raw image and returns a single-channel
-// 2D image corresponding to the red or blue channel using green based interpolation
-void greenBasedRorB(HDRImage &raw, int c, const nanogui::Vector2i &red_offset)
-{
-    // horizontal interpolation
-    parallel_for(red_offset.y(), raw.height(), 2, [&raw,c,&red_offset](int y)
-    {
-        for (int x = red_offset.x() + 1; x < raw.width() - 1; x += 2)
-            raw(x, y)[c] = std::max(0.f, 0.5f * (raw(x - 1, y)[c] + raw(x + 1, y)[c] -
-                                                 raw(x - 1, y)[1] - raw(x + 1, y)[1]) + raw(x, y)[1]);
-    });
-
-    // vertical interpolation
-    parallel_for(red_offset.y() + 1, raw.height() - 1, 2, [&raw,c,&red_offset](int y)
-    {
-        for (int x = red_offset.x(); x < raw.width(); x += 2)
-            raw(x, y)[c] = std::max(0.f, 0.5f * (raw(x, y - 1)[c] + raw(x, y + 1)[c] -
-                                                 raw(x, y - 1)[1] - raw(x, y + 1)[1]) + raw(x, y)[1]);
-    });
-
-    // diagonal interpolation
-    parallel_for(red_offset.y() + 1, raw.height() - 1, 2, [&raw,c,&red_offset](int y)
-    {
-        for (int x = red_offset.x() + 1; x < raw.width() - 1; x += 2)
-            raw(x, y)[c] = std::max(0.f, 0.25f * (raw(x - 1, y - 1)[c] + raw(x + 1, y - 1)[c] +
-                                                  raw(x - 1, y + 1)[c] + raw(x + 1, y + 1)[c] -
-                                                  raw(x - 1, y - 1)[1] - raw(x + 1, y - 1)[1] -
-                                                  raw(x - 1, y + 1)[1] - raw(x + 1, y + 1)[1]) + raw(x, y)[1]);
-    });
 }
 
 } // namespace
